Branding & POJO Stabilization (V13.2): Integrated Phoenix logo and refactored backend to standard POJOs

+++ backend/src/main/java/com/jmrs/gestionparkings/repository/AbonadoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AbonadoRepository extends JpaRepository<Abonado, Long> {
+public interface AbonadoRepository extends JpaRepository<Abonado, Integer> {
     Optional<Abonado> findByDniCif(String dniCif);
     Optional<Abonado> findByEmail(String email);
 }
